Refactored code into seperate files. Refactored discovery code, now focused only on service discovery
