+++ Fleet/Lattice/ServiceManager.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Fleet
+{
+	public class ServiceManager
+	{
+		public ServiceManager ()
+		{
+		}
+	}
+}
+
